it works! now to clean up.
