+++ scriptUtilities/uniqueStr.js
@@ -1,2 +1,2 @@
-const uniqueStr = '(constId: iqNJg1PjpY9@n7y6@NJd)';
-module.exports = { uniqueStr }; 
+const UNIQUE_STR = '(constId: iqNJg1PjpY9@n7y6@NJd)';
+module.exports = { UNIQUE_STR };
